Create interface for UserRepository
