+++ service/src/main/java/com/github/swapbook/repositories/users/FakeUserRepository.java
@@ -1,4 +1,4 @@
-package com.github.swapbook.repositories;
+package com.github.swapbook.repositories.users;
 
 import com.github.swapbook.model.User;
 import org.springframework.stereotype.Component;
@@ -8,13 +8,15 @@ import java.util.List;
 
 
 @Component
-public class UsersRepository {
+public class FakeUserRepository implements UserRepository {
     private List<User> listUsers = new LinkedList<>();
 
+    @Override
     public List<User> getUsers() {
         return listUsers;
     }
 
+    @Override
     public User getUserById(int id){
         return listUsers.stream()
                 .filter(u -> u.getId() == id)
@@ -22,6 +24,7 @@ public class UsersRepository {
                 .orElse(null);
     }
 
+    @Override
     public void addToList(User user){
         listUsers.add(user);
     }
